txDevITelex*: Fix race condition with _rx_buffer
Introduce explicit locking on _rx_buffer to alleviate issues with
_rx_buffer. The pecularity in this case applies only to server operation
and concerns the _rx_buffer modification inside the server's read
method, which lives in one thread. The buffer is normally filled and
gauged from the connection thread. Inside the read method, the
_rx_buffer is iterated over and, in some cases, modified, which is
nonatomic and may lead to buffer corruption.

This should also entail a more robust acknowledge counter calculation.