txDevITelexSrv: Track power state to prevent inadvertent wake-up
Up until now, the code path for terminating a connection as server was
normally only reached directly after a connection was up and the
teleprinter was running: This is why we send an ESC-Z on termination.
The teleprinter driver will disable the latter and MCP will send ESC-ZZ
afterwards to turn off teleprinter power.

With the introduction of incoming ASCII blocking, we might send ESC-Z
whilst the teleprinter is _not_ running (in idle, ZZ state). This leads
to an inadvertent transition from ZZ to Z, turning on teleprinter power
even though it's not needed.

Suppress this ESC-Z by checking if ZZ mode is active currently. To
facilitate this, ESC-ZZ/-Z/-WB/-A/-AA are tracked.